Abstract Panache ORM deployment stuff to panache-common
