--- extensions/panache/hibernate-orm-panache/deployment/src/main/java/io/quarkus/hibernate/orm/panache/deployment/EntityModel.java
@@ -1,19 +1,0 @@
-package io.quarkus.hibernate.orm.panache.deployment;
-
-import java.util.Map;
-
-import org.jboss.jandex.ClassInfo;
-
-public class EntityModel {
-
-    final String name;
-    final String superClassName;
-    final Map<String, EntityField> fields;
-
-    public EntityModel(ClassInfo classInfo, Map<String, EntityField> fields) {
-        this.name = classInfo.name().toString();
-        this.superClassName = classInfo.superName().toString();
-        this.fields = fields;
-    }
-
-}
